sp1_lib::ecdsa

Module affine

The affine point type for SP1. Implementation of an affine point, with acceleration for operations in the context of SP1.

The crate::ecdsa::ProjectivePoint type is mainly used in the ecdsa-core algorithms, however, in some cases, the affine point is required.

Note: When performing curve operations, accelerated crates for SP1 use affine arithmetic instead of projective arithmetic for performance.

Structs

AffinePoint