sp1_lib::ecdsa::affine

Struct AffinePoint

pub struct AffinePoint<C: ECDSACurve> {
    pub inner: C::SP1AffinePoint,
}

Fields
inner: C::SP1AffinePoint
Implementations

impl<C: ECDSACurve> AffinePoint<C>

pub fn from_field_elements_unchecked( x: FieldElement<C>, y: FieldElement<C>, ) -> Self

Create an affine point from the given field elements, without checking if the point is on the curve.

pub fn field_elements(&self) -> (FieldElement<C>, FieldElement<C>)

Get the x and y field elements of the point.

The returned elements are always normalized.

pub fn generator() -> Self

Get the generator point.

pub fn identity() -> Self

Get the identity point.

pub fn is_identity(&self) -> Choice

Check if the point is the identity point.