sp1_lib

Module io

Constants

FD_BLS12_381_INVERSE

The file descriptor through which to access hook_bls12_381_inverse.

FD_BLS12_381_SQRT

The file descriptor through which to access hook_bls12_381_sqrt.

FD_ECRECOVER_HOOK

The file descriptor through which to access hook_ecrecover.

FD_EDDECOMPRESS

The file descriptor through which to access hook_ed_decompress.

FD_FP_INV

The file descriptor through which to access hook_fp_inverse.

FD_FP_SQRT

The file descriptor through which to access hook_fp_sqrt.

FD_HINT

The file descriptor for hints.

FD_PUBLIC_VALUES

The file descriptor for public values.

FD_RSA_MUL_MOD

The file descriptor through which to access hook_rsa_mul_mod.

LOWEST_ALLOWED_FD

The minimum file descriptor.

Functions

commit

Commit a serializable object to the public values stream.

commit_slice

Commit bytes to the public values stream.

hint

Hint a serializable object to the hint stream.

hint_slice

Hint bytes to the hint stream.

read

Read a deserializable object from the input stream.

read_vec

Read a buffer from the input stream. The buffer is read into uninitialized memory.

write

Write the data buf to the file descriptor fd.