sp1_lib::io

Function read_vec

pub fn read_vec() -> Vec<u8> 

Read a buffer from the input stream. The buffer is read into uninitialized memory.

When the bump feature is enabled, the buffer is read into a new buffer allocated by the program.

When the embedded feature is enabled, the buffer is read into the reserved input region.

When there is no allocator selected, the program will fail to compile.

Examples

let data: Vec<u8> = sp1_zkvm::io::read_vec();