sp1_lib

Module utils

Enums

MulAssignError

Errors that can occur during scalar multiplication of an AffinePoint.

WeierstrassPoint

A representation of a point on a Weierstrass curve.

Traits

AffinePoint

WeierstrassAffinePoint

A trait for affine points on Weierstrass curves.

Functions

bytes_to_words_le

Converts a byte array in little endian to a slice of words.

words_to_bytes_le

Converts a slice of words to a byte array in little endian.