sp1_lib::utils

Trait AffinePoint

pub trait AffinePoint<const N: usize>: Clone + Sized {
    const GENERATOR: [u32; N];
    const GENERATOR_T: Self;

    // Required methods
    fn new(limbs: [u32; N]) -> Self;
    fn identity() -> Self;
    fn limbs_ref(&self) -> &[u32; N];
    fn limbs_mut(&mut self) -> &mut [u32; N];
    fn is_identity(&self) -> bool;
    fn add_assign(&mut self, other: &Self);
    fn double(&mut self);

    // Provided methods
    fn from(x: &[u8], y: &[u8]) -> Self { ... }
    fn from_le_bytes(bytes: &[u8]) -> Self { ... }
    fn to_le_bytes(&self) -> Vec<u8>  { ... }
    fn complete_add_assign(&mut self, other: &Self) { ... }
    fn mul_assign(&mut self, scalar: &[u32]) { ... }
    fn multi_scalar_multiplication(
        a_bits_le: &[bool],
        a: Self,
        b_bits_le: &[bool],
        b: Self,
    ) -> Self { ... }

}

Required Associated Constants

const GENERATOR: [u32; N]

๐Deprecated: This const will have the Self type in the next major version.

The generator.

const GENERATOR_T: Self

Required Methods

fn new(limbs: [u32; N]) -> Self

Creates a new AffinePoint from the given limbs.

fn identity() -> Self

Creates a new AffinePoint that corresponds to the identity point.

fn limbs_ref(&self) -> &[u32; N]

Returns a reference to the limbs.

fn limbs_mut(&mut self) -> &mut [u32; N]

Returns a mutable reference to the limbs. If the point is the infinity point, this will panic.

fn is_identity(&self) -> bool

fn add_assign(&mut self, other: &Self)

Adds the given AffinePoint to self.

fn double(&mut self)

Doubles self.

Provided Methods

fn from(x: &[u8], y: &[u8]) -> Self

Creates a new AffinePoint from the given x and y coordinates.

The bytes are the concatenated little endian representations of the coordinates.

fn from_le_bytes(bytes: &[u8]) -> Self

Creates a new AffinePoint from the given bytes in little endian.

fn to_le_bytes(&self) -> Vec<u8>

Creates a new AffinePoint from the given bytes in big endian.

fn complete_add_assign(&mut self, other: &Self)

Adds the given AffinePoint to self. Can be optionally overridden to use a different implementation of addition in multi-scalar multiplication, which is used in secp256k1 recovery.

fn mul_assign(&mut self, scalar: &[u32])

Multiplies self by the given scalar.

fn multi_scalar_multiplication( a_bits_le: &[bool], a: Self, b_bits_le: &[bool], b: Self, ) -> Self

Performs multi-scalar multiplication (MSM) on slices of bit vectors and points. Note: a_bits_le and b_bits_le should be in little endian order.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.